requirements.txt
Pillow==8.3.1
qrcode==7.3
